+++ example/lib/pages/forms/form2.dart
@@ -18,8 +18,8 @@ class _FormData {
 ])
 class Form2Component {
   List<WSelectOption<int>> options = List.generate(20, (idx) => idx + 1)
-      .map((l) => WSelectOption<int>(
-          'Item $l', l, (filter) => 'Item $l'.toLowerCase().contains(filter)))
+      .map((l) => WSelectOption<int>('Item $l', l,
+          (filter) => 'Item $l'.toLowerCase().contains(filter.toLowerCase())))
       .toList();
 
   _FormData data = _FormData();