+++ lib/components/inputs/select/w_select.dart
@@ -6,6 +6,7 @@ import 'package:ng_admin/ng_admin.dart';
 
 class WSelectOption<T> {
   final String label;
+  // make sure the value is unique
   final T value;
   final bool Function(String) compareTo;
 
@@ -14,6 +15,7 @@ class WSelectOption<T> {
 
 @Component(
   selector: 'w-select',
+  styleUrls: ['../input/w_input.css'],
   templateUrl: 'w_select.html',
   providers: [ClassProvider(WInputDecorService)],
   directives: [
@@ -24,7 +26,7 @@ class WSelectOption<T> {
     WInputComponent
   ],
 )
-class WSelectComponent {
+class WSelectComponent implements AfterViewInit {
   @Input('options')
   List<WSelectOption> options = [];
 
@@ -40,6 +42,9 @@ class WSelectComponent {
   @Input('loading')
   set loading(bool flag) => _decorSvc.setLoading(flag);
 
+  @Input('required')
+  bool isRequired = false;
+
   final _valueChange = StreamController<dynamic>();
   @Output()
   Stream<dynamic> get valueChange => _valueChange.stream;
@@ -47,11 +52,10 @@ class WSelectComponent {
   @ViewChild('optionsContainer')
   HtmlElement optionsContainer;
 
-  String filterText;
+  @ViewChild('input')
+  InputElement input;
 
-  List<WSelectOption> get optionsWithFilter => filterText != null
-      ? options.where((element) => element.compareTo(filterText)).toList()
-      : options;
+  String _filterText;
 
   final WInputDecorService _decorSvc;
 
@@ -60,11 +64,17 @@ class WSelectComponent {
     _showingOptions = flag;
     if (!flag) {
       _decorSvc.setFocus(false);
-      filterText = null;
+      _filterText = null;
+    } else {
+      _resetHighlighted();
     }
   }
 
   bool get showOptions => _showingOptions;
+
+  List<WSelectOption> get optionsWithFilter => _filterText != null
+      ? options.where((element) => element.compareTo(_filterText)).toList()
+      : options;
 
   int get selectedItemIdx =>
       value != null ? options.indexWhere((x) => x.value == value) : null;
@@ -72,6 +82,14 @@ class WSelectComponent {
   String get selectedLabel => selectedItemIdx != null && selectedItemIdx >= 0
       ? options[selectedItemIdx].label
       : null;
+
+  WSelectOption _higlightedOption;
+  WSelectOption get highlighted => value != null && _higlightedOption != null
+      ? optionsWithFilter.firstWhere((o) => o.value == _higlightedOption.value,
+          orElse: () => null)
+      : null;
+
+  int get highlightedIdx => optionsWithFilter.indexOf(_higlightedOption);
 
   /// the constructor
   WSelectComponent(this._decorSvc) {
@@ -84,15 +102,48 @@ class WSelectComponent {
         if (ev) _scrollToActive();
       })
       ..onClear.listen((ev) => _valueChange.add(null));
+
+    /// blur on esc
+    document.onKeyDown
+        .where((ev) => ev.keyCode == KeyCode.ESC)
+        .listen((ev) => _onBlur());
+
+    // bind arrow keys
+    document.onKeyDown
+      ..where((ev) =>
+              _showingOptions &&
+              (ev.keyCode == KeyCode.ENTER || ev.keyCode == KeyCode.MAC_ENTER))
+          .listen((ev) => selectOption(_higlightedOption))
+      ..where((ev) =>
+              _showingOptions &&
+              (ev.keyCode == KeyCode.UP || ev.keyCode == KeyCode.DOWN))
+          .listen((KeyboardEvent ev) {
+        ev.preventDefault();
+        ev.keyCode == KeyCode.UP ? _prevHighlight() : _nextHighlight();
+        _scrollToActive();
+      });
+  }
+
+  @override
+  void ngAfterViewInit() {
+    if (isRequired) input.setAttribute('required', '');
+    if (!searchable) input.setAttribute('readonly', '');
+
+    input.onFocus.listen((ev) => _decorSvc.setFocus(true));
+    input.onBlur.listen((ev) => _onBlur());
   }
 
   void selectOption(WSelectOption o) {
     _valueChange.add(o.value);
-    showOptions = false;
+    _onBlur();
   }
 
-  Map<String, bool> getOptionItemClasses(WSelectOption o) =>
-      {'blue': o.value == value};
+  void onInputChange(String v) => _filterText = v.trim();
+
+  Map<String, bool> getOptionItemClasses(WSelectOption o) => {
+        'blue': o.value == value,
+        'bg-light-gray dark-blue': o.value == _higlightedOption?.value
+      };
 
   void _scrollToActive() {
     // get the width of the first child of the option container
@@ -100,6 +151,36 @@ class WSelectComponent {
     int y = optionsContainer.children.isNotEmpty
         ? optionsContainer.children[0].getBoundingClientRect().height.toInt()
         : 50;
-    optionsContainer.scrollTop = (selectedItemIdx ?? 0 * y);
+    optionsContainer.scrollTop = (highlightedIdx * y);
   }
+
+  void _onBlur() {
+    input.blur();
+    showOptions = false;
+    input.value = selectedLabel;
+  }
+
+  void _resetHighlighted() {
+    if (value != null) {
+      _higlightedOption = optionsWithFilter
+          .firstWhere((el) => el.value == value, orElse: () => null);
+    } else if (optionsWithFilter.isNotEmpty) {
+      _higlightedOption = optionsWithFilter[0];
+    }
+  }
+
+  void _nextHighlight() {
+    if (optionsWithFilter.isEmpty) return;
+    int cur = _clamp(highlightedIdx + 1, 0, optionsWithFilter.length);
+    _higlightedOption = optionsWithFilter[cur];
+  }
+
+  void _prevHighlight() {
+    if (optionsWithFilter.isEmpty) return;
+    int cur = _clamp(highlightedIdx - 1, 0, optionsWithFilter.length);
+    _higlightedOption = optionsWithFilter[cur];
+  }
+
+  int _clamp(int idx, int min, int max) =>
+      idx < min ? max - 1 : idx > max - 1 ? min : idx;
 }